Refactored graph view into separate view and graph class.
Connections and Nodes are now part of a CloudGraph class that manages all the neccessary things to display a graph.
Navigation of the main graph was moved from the main UI to a UIGraph (Which is mostly used to display the big main graph and control the camera (and spaceship).
This makes it much easier to generate new sub-graphs and have the navigation separated from the actual graph view and generation logic.