Made calculation of graph layout thread-safe and implemented basic multi-threading.
The Node Structure now represents the Cloud ancestor structure making many translations local instead of global.
The graph layout calculation is now done in predefined groups making iterations times much smaller.
In this commit, the final layout is not as nice as before - More layouting iterations have to be applied.