[ckcore] Introduce the TaskSurpassBehaviour wait (#509)
A TaskDescriptor can now use surpass behaviour wait.
If a task is started and there is already a task with the same descriptor already running, the task is added to a stating queue. Once the existing task is finished, the next task will be started directly.

Note: There can be only one item in the waiting queue per task descriptor. If one task is already running, one already in the waiting queue and another one is triggered, the waiting queue does not change.

Fixes #172